pyproject.toml
[project]
name = "base-fastapi-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "cst-lsp>=0.1.3",
    "fastapi>=0.115.12",
    "greenlet>=3.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pytest-async-sqlalchemy>=0.2.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-dotenv>=0.5.2",
    "pytest>=8.4.0",
    "python-multipart>=0.0.20",
    "ruff>=0.12.0",
    "sentry-sdk>=2.27.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]

[tool.ruff]
line-length = 120
src = ["."]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "ERA",  # found commented-out code
    "D401", # check docstring format
]
unfixable = ["F401"] # Prevent Ruff from auto-fixing unused imports
ignore = [
    "B008", # Ignore depends in function calls
    "E731", # Allow to use the lambda functions
]
exclude = [".venv", ".git", "__pycache__", "build", "dist"]

[tool.uv.workspace]
members = ["venv"]

[tool.pyright]
# Delete if you dont have mistakes
executionEnvironments = [{ root = "." }]
typeCheckingMode = "standard"
venv = ".venv"
venvPath = "."

